Factor out estimateScoreWASM function
